优化代码

+++ Assets/FunkyCode/SmartLighting2D/Scripts/Rendering/Day/Pass.cs
@@ -116,7 +116,7 @@ namespace FunkyCode.Rendering.Day
                 return(false);
             }
 
-            layerId = (int)slayer.id;
+            layerId = slayer.id;
             layer = slayer;
 
             this.camera = camera;
